NativeBinder: Correctly bind empty strings and blobs
Calling .c_str() on an empty string or .data() on an empty vector is
allowed to return a nullptr. SQLite interprets passing a nullptr to
sqlite3_bind_text or sqlite3_bind_blob as if sqlite3_bind_null had been
called. So while we wanted to bind an empty string to the statement, we
bind NULL, which is not the same thing ("" != NULL).

This misbehavior is fixed by replacing an incoming nullptr with a non-
null pointer if the size of the incoming data is zero.